dbeaver/dbeaver#36388 Moved removed project listener to workbench context listener
